+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-30T07:21:16.506Z",
+    "2026-06-30T08:15:15.305Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {