+++ packages/orq-rc/RELEASES.md
@@ -10379,3 +10379,13 @@ Based on:
 - [typescript v4.11.0-rc.59] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.59 - packages/orq-rc
+
+## 2026-06-30 08:19:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.60] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.60 - packages/orq-rc